% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_map.R
\name{plot_route_map}
\alias{plot_route_map}
\title{Plot Running Route from GPX}
\usage{
plot_route_map(gpx_file)
}
\arguments{
\item{gpx_file}{}
}
\value{
A tmap map object displaying the running route.
}
\description{
This function reads a .gpx file and plots the running route using the 'tmap' package and OpenStreetMap tiles.
}
\examples{
\dontrun{
plot_route_map("example.gpx")
}
}
